+++ external/price-scrapper/app/database/__init__.py
@@ -1,3 +1,3 @@
-from app.database.con import *
+from app.database.conn import *
 from app.database.prices import *
 from app.database.products import *
